fix!: remove FormatSync and FormatWrite
Use `Format` for everything now after rewrite. No need to have a command
that formats the buffer synchronously. Also no need for `FormatWrite`
anymore after fixing the biggest painpoint with self.is_formatting and
`vim.wait`. It should always be set to false now since we should be able
to catch errors even in an async context. Also, `VimLeavePre` as an even
is working with formatting from the command line