fix(format_on_save)!: autocmd on _all_ buffers if format_on_save enabled
Previously, I only created an autocmd for format_on_save if the filetype
of the buffer had a config for formatting. This made it not work with
format_on_save for injections if the filetype of the buffer didn't have
any formatters available.

Pros: Now it should work all the time

Cons: It will now parse with treesitter on _every_ save in _all_
filetypes if this option is enabled.

If you would like the previous behaviour, you could create an autocmd
for yourself with something like:

```lua
vim.api.nvim_create_autocmd('FileType', {
    pattern = { "lua", "rust", "c" }, -- All your filetypes here
    group = vim.api.nvim_create_augroup('nvim-formatter_on_save_ft', { clear = true }),
    callback = function()
        local bufnr = vim.api.nvim_get_current_buf()
        vim.api.nvim_create_autocmd('BufWritePost', {
            buffer = bufnr,
            group = vim.api.nvim_create_augroup(
                string.format('nvim-formatter_on_save_buf_%s', bufnr),
                { clear = true }
            ),
            callback = function()
                -- Add whatever logic you want here to potentially not
                -- format
                vim.cmd.Format()
            end,
        })
    end,
})
````